requirements.txt
gunicorn==21.2
python-telegram-bot==20.8